+++ components/ProfileSidebar.vue
@@ -54,12 +54,12 @@ const props = defineProps({
     type: Array,
     default: () => [
       { name: "My Account", action: "account", route: "/account" },
-      { name: "Orders", action: "orders", route: "/table-orders" },
-      { name: "Invoices", action: "invoices", route: "/invoices" },
-      { name: "Credits", action: "credits" },
-      { name: "Shipments", action: "shipments" },
-      { name: "Back Orders", action: "back-orders" },
-      { name: "Complaints", action: "complaints" },
+      { name: "Orders", action: "orders", route: "/account/orders" },
+      { name: "Invoices", action: "invoices", route: "/account/invoices" },
+      { name: "Credits", action: "credits", route: "/account/credits" },
+      { name: "Shipments", action: "shipments", route: "/account/shipments" },
+      { name: "Back Orders", action: "back-orders", comingSoon: true },
+      { name: "Complaints", action: "complaints", comingSoon: true },
     ],
   },
   showOrderUpload: {
