+++ components/SearchForm.vue
@@ -13,7 +13,7 @@
       >
         <label
           :for="field.key"
-          class="uppercase text-[#3D3935] w-[120px] text-base"
+          class="uppercase text-[#3D3935] w-[120px] text-base font-light"
         >
           {{ field.label }}
         </label>
@@ -21,7 +21,7 @@
           v-model="formData[field.key]"
           :id="field.key"
           :type="field.type"
-          class="flex-1 rounded-full px-4 py-2 text-[#6F6259] border border-[#C9C7C5] focus:outline-none uppercase"
+          class="flex-1 rounded-[66px] px-4 py-2 text-[var(--Ash-Grey, #6F6259)] font-light border border-[var(--Line-Strong, #C9C7C5)] focus:outline-none uppercase"
           :placeholder="field.placeholder || ''"
         />
       </div>
